+++ CHANGELOG.md
@@ -146,3 +146,4 @@
 * Rename `|.:.|` type alias to `=!?!=`
 * Define `Adj.Program.Controlflow` module
 * Define `Adj.Program.Controlflow.Implementation` module
+* Define `Adj.Program.Controlflow.Implementation.State` module